+++ Runtime/Scripts/Components/ExtendedBehaviour.cs
@@ -1,29 +1,25 @@
-using System.Collections.Generic;
-using System.Linq;
+using System;
+using ParkMinPackages.Foundation.Interfaces;
 using R3;
 using UnityEngine;
 
 namespace ParkMinPackages.Foundation.Components
 {
-	public abstract class ExtendedBehaviour : MonoBehaviour
+	public abstract class ExtendedBehaviour : MonoBehaviour, IDisposable
 	{
-		//Statics-------------------------------------------------------------------------------------------
-		protected enum R3UpdateMethod
-		{
-			EarlyUpdateR3,
-			FixedUpdateR3,
-			PreUpdateR3,
-			UpdateR3,
-			PreLateUpdateR3,
-			PostLateUpdateR3,
-			TimeUpdateR3,
-			PostFixedUpdateR3
-		}
-
 		//Public Methods-------------------------------------------------------------------------------------------
 		public bool IsStarted
 		{
 			get { return _isStarted; }
+		}
+		public void Dispose() {
+			if (_isDisposed) {
+				return;
+			}
+
+			_isDisposed = true;
+			enabled = false;
+			Destroy(this);
 		}
 		//Public Properties----------------------------------------------------------------------------------------
 
@@ -36,26 +32,40 @@ namespace ParkMinPackages.Foundation.Components
 			}
 		}
 		protected virtual void OnReady() {
-			HashSet<R3UpdateMethod> r3UpdateMethods = DefineR3UpdateMethod()?.ToHashSet();
-			if (r3UpdateMethods != null && 1 <= r3UpdateMethods.Count) {
-				_r3UpdateMethodsDisposable?.Dispose();
-				_r3UpdateMethodsDisposable = new CompositeDisposable();
-				if (r3UpdateMethods.Contains(R3UpdateMethod.EarlyUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.EarlyUpdate).Subscribe(unit => { EarlyUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.FixedUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.FixedUpdate).Subscribe(unit => { FixedUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.PreUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.PreUpdate).Subscribe(unit => { PreUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.UpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.Update).Subscribe(unit => { UpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.PreLateUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.PreLateUpdate).Subscribe(unit => { PreLateUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.PostLateUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.PostLateUpdate).Subscribe(unit => { PostLateUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.TimeUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.TimeUpdate).Subscribe(unit => { TimeUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
-				if (r3UpdateMethods.Contains(R3UpdateMethod.PostFixedUpdateR3))
-					Observable.EveryUpdate(UnityFrameProvider.PostFixedUpdate).Subscribe(unit => { PostFixedUpdateR3(); }).AddTo(_r3UpdateMethodsDisposable);
+			_r3UpdateMethodsDisposable?.Dispose();
+			_r3UpdateMethodsDisposable = null;
+
+			if (this is IR3EarlyUpdatable earlyUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.EarlyUpdate).Subscribe(_ => earlyUpdatable.R3EarlyUpdate()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3FixedUpdatable fixedUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.FixedUpdate).Subscribe(_ => fixedUpdatable.R3FixedUpdate()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3PreUpdatable preUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.PreUpdate).Subscribe(_ => preUpdatable.R3PreUpdate()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3Updatable updatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.Update).Subscribe(_ => updatable.R3Update()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3PreLateUpdatable preLateUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.PreLateUpdate).Subscribe(_ => preLateUpdatable.R3PreLateUpdate()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3PostLateUpdatable postLateUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.PostLateUpdate).Subscribe(_ => postLateUpdatable.R3PostLateUpdate()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3TimeUpdatable timeUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.TimeUpdate).Subscribe(_ => timeUpdatable.R3TimeUpdate()).AddTo(_r3UpdateMethodsDisposable);
+			}
+			if (this is IR3PostFixedUpdatable postFixedUpdatable) {
+				_r3UpdateMethodsDisposable ??= new CompositeDisposable();
+				Observable.EveryUpdate(UnityFrameProvider.PostFixedUpdate).Subscribe(_ => postFixedUpdatable.R3PostFixedUpdate()).AddTo(_r3UpdateMethodsDisposable);
 			}
 		}
 		protected virtual void Start() {
@@ -64,6 +74,7 @@ namespace ParkMinPackages.Foundation.Components
 		}
 		protected virtual void OnDisable() {
 			_r3UpdateMethodsDisposable?.Dispose();
+			_r3UpdateMethodsDisposable = null;
 			if (Application.exitCancellationToken.IsCancellationRequested == false) {
 				OnDisableDuringRuntime();
 			}
@@ -74,20 +85,9 @@ namespace ParkMinPackages.Foundation.Components
 			}
 		}
 		//Internals---------------------------------------------------------------------------------------
+		bool _isDisposed;
 		bool _isStarted;
 		CompositeDisposable _r3UpdateMethodsDisposable;
-
-		protected virtual R3UpdateMethod[] DefineR3UpdateMethod() {
-			return new R3UpdateMethod[] { };
-		}
-		protected virtual void EarlyUpdateR3() { }
-		protected virtual void FixedUpdateR3() { }
-		protected virtual void PreUpdateR3() { }
-		protected virtual void UpdateR3() { }
-		protected virtual void PreLateUpdateR3() { }
-		protected virtual void PostLateUpdateR3() { }
-		protected virtual void TimeUpdateR3() { }
-		protected virtual void PostFixedUpdateR3() { }
 
 		protected virtual void OnDisableDuringRuntime() { }
 		protected virtual void OnDestroyDuringRuntime() { }