+++ Runtime/Scripts/Extensions/ComponentExtensions.cs
@@ -1,9 +1,18 @@
+using System;
 using UnityEngine;
 
 namespace ParkMinPackages.Foundation.Extensions
 {
 	public static class ComponentExtensions
 	{
+		public static Component AddComponent(this Component component, Type componentType) {
+			return component.gameObject.AddComponent(componentType);
+		}
+
+		public static T AddComponent<T>(this Component component) where T : Component {
+			return component.gameObject.AddComponent<T>();
+		}
+
 		public static bool IsSceneObject(this Component component) {
 			if (component == null)
 				return false;
@@ -15,7 +24,7 @@ namespace ParkMinPackages.Foundation.Extensions
 			if (component.TryGetComponent<T>(out T t))
 				return t;
 			else
-				return component.gameObject.AddComponent<T>();
+				return component.AddComponent<T>();
 		}
 	}
 }