+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.0.0] - 2026-08-22
+
+### Breaking Changes
+- Replaced `ExtendedBehaviour` update-method declarations and virtual callbacks with opt-in R3 update interfaces.
+
+### Added
+- Added R3 update interfaces for each supported Unity player-loop phase.
+- Added `IDisposable` support to `ExtendedBehaviour` for disabling and destroying an attached component.
+- Added `Component.AddComponent` extensions for generic and runtime component types.
+
 ## [6.0.0] - 2026-08-20
 
 ### Changed